requirements.txt
openai
SpeechRecognition
pyaudio
python-dotenv
soundfile
sounddevice
asyncio
setuptools
numpy
